feat(DeleteExpenseUseCase): check if there are any expenses so you can delete them.

+++ src/modules/expenses/use-cases/DeleteExpenseUseCase.ts
@@ -1,4 +1,5 @@
 import { IExpenseRepository } from "../../../interfaces/IExpenseRepository";
+import UseCaseError from "../../../interfaces/UseCaseError";
 
 class DeleteExpenseUseCase {
   private expenseRepository: IExpenseRepository;
@@ -8,6 +9,16 @@ class DeleteExpenseUseCase {
   }
 
   async execute(expenseId: number): Promise<void> {
+    const existingExpense =
+      await this.expenseRepository.findExpenseById(expenseId);
+
+    if (!existingExpense) {
+      throw <UseCaseError>{
+        message: "This expense does not exist.",
+        errorType: "EXPENSE_NOT_FOUND",
+      };
+    }
+
     return await this.expenseRepository.deleteExpense(expenseId);
   }
 }
